Refactor sdk/compiler/handlers to split into smaller files
