+++ .rubocop_todo.yml
@@ -39,10 +39,6 @@ Metrics/CyclomaticComplexity:
 Metrics/MethodLength:
   Max: 39
 
-# Offense count: 1
-Metrics/PerceivedComplexity:
-  Max: 11
-
 # Offense count: 15
 Style/Documentation:
   Exclude:
